Final version of menu animation
Finally I fixed all bugs in a menu animation
and menu component in general. In this version
we animate our menu by mathematical calculations.
From custom hooks we get the screen and the ref element
width and height, then we divide it and our result
is a scale we have to use to transform our menu.